[Rename] startGames -> startGame 폴더명 변경
