+++ app/speedgame/start-game/[room]/page.tsx
@@ -1,4 +1,4 @@
-import StartSpeedGame from '@/components/screen/startGames/SpeedGame';
+import StartSpeedGame from '@/components/screen/startGame/SpeedGame';
 
 export default function StartGame() {
 	return <StartSpeedGame />;
