+++ components/screen/startGame/MiniNetworking.tsx
