+++ components/screen/startGame/MoodCheckIn.tsx
