+++ components/screen/startGame/SpeedGame.tsx
