+++ components/screen/startGame/ThankCircle.tsx
